src/python-server/requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn
pydantic
starlette

# Environment and configuration
python-dotenv

# Windows COM automation for PowerPoint
pywin32
pywin32-ctypes

# Image processing
pillow

# Vector database and embeddings
pinecone
voyageai

# Utilities
requests
certifi
typing_extensions
